Discard bytes on STDIN when reading DCS
This prevents errors when TerminalExtensions is required within an `atreplinit` function in juliarc.jl, and an impatient user starts typing before the prompt is ready. It would be really nice to save these bytes in a buffer and put them back into STDIN once TerminalExtensions has done its initialization, but I cannot figure out how to get the REPL to respect these bytes. In any case, I find this to be much more usable -- I would rather have my input discarded and TerminalExtensions to work than get jibberish on the first prompt with a failure.